Version increments and reorg
…while investigating #2
